% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoLQA.R
\name{lassoLQA}
\alias{lassoLQA}
\title{Select variables via lasso in linear regression.}
\usage{
lassoLQA(x, y)
}
\arguments{
\item{x}{A matrix.}

\item{y}{A vector.}
}
\value{
The result of selected variables.
}
\description{
Select variables via lasso in linear regression.
}
\examples{
lassoLQA(matrix(rnorm(90),30,3), matrix(rnorm(90),30,3)\%*\%c(2,1,0)+rnorm(30,0,4))
lassoLQA(matrix(rnorm(1200),200,6), matrix(rnorm(1200),200,6)\%*\%c(3,2,1,0,0,0)+rnorm(200,0,4))
}
